+++ src/constants/banner-data.ts
