+++ src/constants/space-category-data.ts
